BaseRepo

+++ common/common-dao/src/main/java/com/meysam/common/dao/BaseRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.NoRepositoryBean;
 import java.math.BigInteger;
 
 @NoRepositoryBean
-public interface BaseRepository<ENTITY_TYPE> extends JpaRepository<ENTITY_TYPE, BigInteger>, CrudRepository<ENTITY_TYPE, BigInteger>, QuerydslPredicateExecutor<ENTITY_TYPE> {
+public interface BaseRepository<ENTITY_TYPE> extends JpaRepository<ENTITY_TYPE, Long>, CrudRepository<ENTITY_TYPE, Long>, QuerydslPredicateExecutor<ENTITY_TYPE> {
 }
